+++ tools/pushToDockerHub.sh
@@ -18,7 +18,7 @@ fi
 
 declare IMAGE_ID=$(docker images -q ${IMAGE_NAME}:${TAG})
 if [ -z "$IMAGE_ID" ]; then
-    echo "No Docker image $IMAGE_NAME:$TAG has been built, do 'gradle distDocker' first!" >&2
+    echo "No Docker image $IMAGE_NAME:$TAG has been built, do 'gradle buildDockerImage' first!" >&2
     exit 3
 fi